Now double is supported.

+++ src/database/HSMDatabase.AccessManager/DatabaseEntities/VisualEntity/DashboardPanelEntity.cs
@@ -31,9 +31,9 @@ namespace HSMDatabase.AccessManager.DatabaseEntities.VisualEntity
 
         public bool ShowLegend { get; init; }
         
-        public int MaxY { get; set; }
+        public double MaxY { get; set; }
         
-        public int MinY { get; set; }
+        public double MinY { get; set; }
         
         public bool AutoScale { get; set; }
     }
